replace timer from GameStep to Game
